Adding a chunk_memset variant that is VBMI aware
This version is marginally faster on CPUs that support VBMI such as
Icelake and beyond. It has a more flexible cross lane permute instruction
for byte sized values that is used when the copy distance >= 16. It also
has slightly more optimal behavior for when dist < 16 by doing independent
loads and shuffles.